requirements.txt
crewai==0.35.8
crewai-tools==0.3.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain-google-genai==1.0.3
pypdf==4.2.0